Minor refactoring to frame work to be consistence between acceptance and functional
